Add files via upload
In this project we processed wind data, observed effect of different capabilities for wind turbines, techno-economically evaluated and proposed a Wind Energy project for a given isolated system. Included non-requested optimization of spatial placement of wind turbines in the wind farm, utilizing basic scripting and the E.A.SY software. The report includes code snippets in FORTRAN and MATLAB, some scripting along with the configurations of EA.SY (optimization software based on Evolutionary Algorithms, developed by the PCOpt Unit in NTUA).